+++ CHANGELOG.md
@@ -24,6 +24,10 @@
 - Adds links to Artwork view for FAQ and Ask a Specialist - kierangillen
 - Adds auction extra links in Artwork view - kierangillen
 - Fix shows artwork filter pickers always reseting to default value - zephraph
+- Adds AuctionPrice to CommercialInformation to render Auction price data - kierangillen
+
+### Master
+
 - Adds Commercial Bid Buttons - kierangillen
 - Adds VAT included in price to Artwork view - kierangillen
 - Replace slug keys with ids - zephraph